requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.3
flask-socketio==5.3.6
flask-cors==4.0.0
pymysql==1.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
eventlet==0.33.3